refactor: extract `AgentInfos`
